+++ example/example.dart
@@ -6,17 +6,50 @@ void main() {
   /// USD currency uses 2 decimal places.
   ///
   final usd = Currency.create('USD', 2);
+  final aud = Currency.create('USD', 2);
 
   ///
   /// Create a money which stores $USD 10.00
   ///
-  /// Note: we use the minor unit (e.g. cents) when passing in the 
+  /// Note: we use the minor unit (e.g. cents) when passing in the
   ///   monetary value.
   /// So $10.00 is 1000 cents.
   ///
   Money costPrice = Money.fromInt(1000, usd);
   print(costPrice.toString());
   // > $10.00
+
+  ///
+  /// Create a [Money] instance from a String
+  /// using [Currency.fromString]
+  /// The [Currency] of salePrice is USD.
+  ///
+  Money salePrice = usd.fromString("\$10.50");
+  print(salePrice.format("SCC 0.0"));
+  // > $US 10.50
+
+  ///
+  /// Create a [Money] instance from a String
+  /// using [Money.fromString]
+  ///
+  Money taxPrice = Money.fromString("\$1.50", usd);
+  print(taxPrice.format("CC 0.0 S"));
+  // > US 1.50 $
+
+  ///
+  /// Create a [Money] instance from a String
+  /// with an embedded Currency Code
+  /// using [Currencies.fromString]
+  ///
+  Currencies.register(usd);
+  Currencies.register(aud);
+  Money cheapIPhone = Currencies.fromString("\$USD1500.0", "SCCC0.0");
+  print(cheapIPhone.format("SCC0.0"));
+  // > $US1500.00
+
+  Money expensiveIPhone = Currencies.fromString("\$AUD2000.0", "SCCC0.0");
+  print(expensiveIPhone.format("SCC0.0"));
+  // > $AUD2000.00
 
   ///
   /// Do some maths
@@ -32,7 +65,7 @@ void main() {
   ///
   /// Do some custom formatting of the ouput
   /// S - the symbol e.g. $
-  /// CC - first two digits of the currency code provided when creating 
+  /// CC - first two digits of the currency code provided when creating
   ///     the currency.
   /// # - a digit if required
   /// 0 - a digit or the zero character as padding.
@@ -44,11 +77,11 @@ void main() {
   ///    when calling [Money.toString()]
   ///
   /// JPY - code for japenese yen.
-  /// 0 - the number of minor units (e.g cents) used by the currency. 
+  /// 0 - the number of minor units (e.g cents) used by the currency.
   ///     The yen has no minor units.
   /// ¥ - currency symbol for the yen
-  /// S0 - the default pattern for [Money.toString()]. 
-  ///      S output the symbol. 
+  /// S0 - the default pattern for [Money.toString()].
+  ///      S output the symbol.
   ///      0 - force at least a single digit in the output.
   ///
   final Currency jpy = Currency.create('JPY', 0, symbol: '¥', pattern: 'S0');
@@ -58,10 +91,10 @@ void main() {
 
   ///
   /// Define a currency that has inverted separators.
-  /// i.e. The USD uses '.' for the integer/fractional separator 
+  /// i.e. The USD uses '.' for the integer/fractional separator
   ///      and ',' for the thousands separator.
   ///      -> 1,000.00
-  /// The EURO use ',' for the integer/fractional separator 
+  /// The EURO use ',' for the integer/fractional separator
   ///      and '.' for the thousands separator.
   ///      -> 1.000,00
   ///
@@ -100,18 +133,17 @@ void main() {
   print(euroCostPrice.format("###.###,#0 CC"));
   // > 100.345,30 EU
 
-  /// 
-  /// Make the currencies available globally by registering them 
+  ///
+  /// Make the currencies available globally by registering them
   ///     with the [Currencies] singleton factory.
-  /// 
-  Currencies().register(usd);
-  Currencies().register(euro);
-  Currencies().register(jpy);
+  ///
+  Currencies.register(usd);
+  Currencies.register(euro);
+  Currencies.register(jpy);
 
-   
   // use a registered currency by finding it in the registry using
   // the currency code that the currency was created with.
-  Currency usDollar = Currencies().find("USD");
+  Currency usDollar = Currencies.find("USD");
 
   Money invoicePrice = Money.fromInt(1000, usDollar);