+++ lib/src/currencies.dart
@@ -22,7 +22,11 @@
  * THE SOFTWARE.
  */
 
+import 'package:money2/src/pattern_decoder.dart';
+
 import 'currency.dart';
+import 'money.dart';
+import 'money_data.dart';
 
 /// A factory for registering and accessing [Currency] instances.
 ///
@@ -37,27 +41,55 @@ import 'currency.dart';
 /// them as needed.
 
 class Currencies {
-  static Currencies _self = Currencies._internal();
+  // static Currencies _self = Currencies._internal();
 
-  final Map<String, Currency> _directory;
-
-  factory Currencies() {
-    _self = Currencies._internal();
-    return _self;
-  }
-
-  Currencies._internal() : _directory = Map();
+  static final Map<String, Currency> _directory = Map();
 
   /// Register a Currency
-  void register(Currency currency) {
+  static void register(Currency currency) {
     _directory[currency.code] = currency;
   }
 
   /// Register a list of currencies.
-  void registerList(Iterable<Currency> currencies) {
+  static void registerList(Iterable<Currency> currencies) {
     currencies.forEach((currency) {
       _directory[currency.code] = currency;
     });
+  }
+
+  ///
+  /// Parses a string containing a money amount including a currency code.
+  /// Provided the passed currency code belongs to a [Currency]
+  /// that has been registered via [Currencies.register] or
+  /// [Currencies.registerList] thenthis method will return a
+  /// [Money] instance of that [Currency] type.
+  ///
+  /// A [MoneyParseException] is thrown if the [monetaryAmount]
+  /// doesn't match the [pattern].
+  ///
+  /// An [UnknownCurrencyException] is thrown if the [monetaryAmount]
+  /// does not contain a known currency.
+  ///
+  static Money fromString(String monetaryAmount, String pattern) {
+    int codeLength = _getCodeLength(pattern);
+
+    if (codeLength < 2) {
+      throw MoneyParseException(
+          "The Country Code length (e.g. CC) must be at least 2 characters long");
+    }
+
+    String code = _extractCode(monetaryAmount, codeLength);
+
+    Currency currency = find(code);
+
+    if (currency == null) {
+      throw UnknownCurrencyException(code);
+    }
+
+    PatternDecoder decoder = PatternDecoder(currency, pattern);
+    MoneyData moneyData = decoder.decode(monetaryAmount);
+
+    return Money.fromBigInt(moneyData.minorUnits, currency);
   }
 
   /* Protocol *****************************************************************/
@@ -65,7 +97,42 @@ class Currencies {
   /// Searches the list of registered [Currency]s.
   ///
   /// Returns the [Currency] that matches [code] or `null` if no matching [code] is found.
-  Currency find(String code) {
+  static Currency find(String code) {
     return _directory[code];
+  }
+
+  /// Counts the number of 'C' in a pattern
+  static int _getCodeLength(String pattern) {
+    int count = 0;
+    for (int i = 0; i < pattern.length; i++) {
+      if (pattern[i] == 'C') count++;
+    }
+    return count;
+  }
+
+  static String _extractCode(String monetaryValue, int codeLength) {
+    RegExp regEx = RegExp("[A-Za-z]" * codeLength);
+
+    var matches = regEx.allMatches(monetaryValue);
+    if (matches.isEmpty) {
+      throw MoneyParseException(
+          "No currency code found in the pattern: $monetaryValue");
+    }
+
+    if (matches.length > 1) {
+      throw MoneyParseException(
+          "More than one currency code found in the pattern: $monetaryValue");
+    }
+
+    return monetaryValue.substring(matches.first.start, matches.first.end);
+  }
+}
+
+class UnknownCurrencyException implements Exception {
+  String code;
+  UnknownCurrencyException(this.code);
+
+  String toString() {
+    return "An unknown currency '$code' was passed. Register the currency via [Currencies.register()] and try again.";
   }
 }